selenium-tests/requirements.txt
selenium>=4.10.0
pytest>=7.0.0
pytest-xdist
allure-pytest
